fixup! merge Writer and StreamWriter to FileStoreWriter
